Extract reporting from tasks.py
