+++ exasol/toolbox/nox/tasks.py
@@ -21,11 +21,6 @@ import argparse
 import nox
 from nox import Session
 
-from exasol.toolbox.metrics import (
-    Format,
-    create_report,
-    format_report,
-)
 from exasol.toolbox.nox._documentation import (
     build_docs,
     clean_docs,
@@ -42,6 +37,7 @@ from exasol.toolbox.nox._lint import (
     lint,
     type_check,
 )
+from exasol.toolbox.nox._metrics import report
 from exasol.toolbox.nox._release import prepare_release
 from exasol.toolbox.nox._shared import (
     Mode,
@@ -69,52 +65,3 @@ def check(session: Session) -> None:
     _pylint(session, py_files)
     _type_check(session, py_files)
     _coverage(session, PROJECT_CONFIG, context)
-
-
-@nox.session(name="report", python=False)
-def report(session: Session) -> None:
-    """
-    Collects and generates metrics summary for the workspace
-
-    Attention:
-
-        Pre-requisites:
-
-        * Make sure you remove old and outdated artifacts
-            - e.g. by running one of the following commands
-                * :code:`git clean -xdf`
-                * :code:`rm .coverage .lint.txt`
-
-        * Run the following targets:
-            - :code:`nox -s coverage`
-            - :code:`nox -s lint`
-    """
-    formats = tuple(fmt.name.lower() for fmt in Format)
-    usage = "nox -s report -- [options]"
-    parser = argparse.ArgumentParser(
-        description="Generates status report for the project", usage=usage
-    )
-    parser.add_argument(
-        "-f",
-        "--format",
-        type=str,
-        default=formats[0],
-        help="Output format to produce.",
-        choices=formats,
-    )
-    required_files = (
-        PROJECT_CONFIG.root / ".coverage",
-        PROJECT_CONFIG.root / ".lint.txt",
-    )
-    if not all(file.exists() for file in required_files):
-        session.error(
-            "Please make sure you run the `coverage` and the `lint` target first"
-        )
-    sha1 = str(
-        session.run("git", "rev-parse", "HEAD", external=True, silent=True)
-    ).strip()
-    args: argparse.Namespace = parser.parse_args(args=session.posargs)
-    project_report = create_report(commit=sha1)
-    fmt = Format.from_string(args.format)
-
-    print(format_report(project_report, fmt))
